WIP: use piecewise lux calculation example from spec

+++ include/I2c/Devices/Tsl2561LuxMeter/Tsl2561LuxValues.h
@@ -7,8 +7,8 @@ namespace I2c
 
 struct Tsl2561LuxValues
 {
-		uint16_t irSpectrumLux;
-		uint16_t visibleSpectrumLux;
+		uint32_t irSpectrumLux;
+		uint32_t visibleSpectrumLux;
 };
 
 } // namespace I2c
